terminal_app/requirements.txt
pyscard
pycryptodome